backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.8
pillow==12.1.1
PyJWT==2.11.0
sqlparse==0.5.5
